[pulse][taint] Improve reporting of data flows from taint sources
Summary:
Improve reporting of "sensitive data flows" for flows which are interprocedural.
In particular, try to report all flows from known taint sources to arguments of procedure calls.

This allows Infer to report data flows in `SensitiveDataFlow.m` from `__infer_taint_source` all the way to `might_be_a_sink` - previously it could only get as far as `mutate_then_consume`.

For example:
  function foo(obj) {
    bar(obj);
  }

  function baz() {
    obj = taint_source();
    foo(obj);
  }
Here we want to report a sensitive data flow from `taint_source` in `baz` to the call to `bar` in `foo`.
However, the summary for `foo` does not contain sufficient information to conclude that the argument to `foo` flows
to `bar` - when we analyse the call to `foo`, we cannot report the flow.
This commit adds to the summary for `foo`: the new attribute makes clear that the argument to `foo` will flow to `bar`.

More specifically:
  - Augment the `MustNotBeTainted` attribute to keep track of "taint procedures" (as well as known taint sinks):
    these are expressed as a set of `{proc_name; argument_index; trace}`.
    When tainted, an address with this attribute could taint all the `argument_index`/`proc_name` pairs in the set
    via the corresponding `trace` - i.e. taint flows to the address may taint `argument_index` of `procname`.
    Another way to see this: all the `MustNotBeTainted` attributes with sets containing the procedure `foo`
    represent ways of tainting the input to `foo`.
  - Add this attribute to Pulse state during function calls: mark each argument as a way of tainting input to the function.
  - Report `SensitiveDataFlow`s during function calls: match up incoming tainted flows to "taint procedures" reached
    when calling the function.

Reviewed By: dulmarod

Differential Revision: D37348114

fbshipit-source-id: 71b8362763e04b45a9008c616859f64d3a9707ae