[pulse][taint] Add a flag to specify a taint config file
Summary:
Add the flag `--pulse-taint-config <path>`, allowing users to specify JSON file(s) which will contain taint configuration.
In particular, these can define the following keys in exactly the same way as a `.inferconfig`:
  pulse-taint-policies
  pulse-taint-sources
  pulse-taint-sanitizers
  pulse-taint-sinks

When the flag is specified, the passed JSONs will be concatenated with the result of the existing `.inferconfig`/commandline
processing. Therefore behaviour is as before if the flag is not specified.

This new option is exercised by using it in Objective-C taint analysis tests: split the existing `.inferconfig` into three
separate taint configurations, one common to tests in both `objc/pulse/taint` and `objc/pulse-data-lineage`, and two
additional ones specific to each directory.

Reviewed By: skcho

Differential Revision: D37550804

fbshipit-source-id: 8977e1807717b66cf12211e81e76506ad6bd327f