[racerd] eliminate dep cycles from attributes
Summary:
In order to silence reports on (dynamically typed/initialised) synchronised containers, RacerD processes the summary of the class initialiser and of every constructor, which contain information on container type initialisation.

Because this happens at the procedure analysis level of every method, this creates a non-call dependency which sometimes leads to dependency cycles, eg in Enums, the class initialiser calls a constructor. These cycles introduce non-determinism due to the (non)-handling of recursion.

This diff removes that analysis dependency by accumulating all containers accesses in a method summary (as if there are no synchronised container) and then at the file analysis level filters that summary given those of the class initialiser and constructors (which have already been analysed at the procedure level).

Reviewed By: skcho

Differential Revision: D33622613

fbshipit-source-id: 60ae10fa33d4c90550ef97f38018c6a501197f6d