[lineage] A new way of removing temps
Summary:
Previously, the reduction worked by replacing a path
  NT_A -> T_1 -> ... -> T_n -> NT_B
by an edge
  NT_A -> NT_B
where NT stands for a non-temporary piece of data, and T for a temporary
(introduced by the frontend).

This reduced the number of nodes in lineage by as much as possible: all
nodes corresponding to temporaries disappeared. However, the number of
edges could have increased greatly.

The other problem was that we had to come up with some heuristics about
what information from the whole path we retain on the single edge that
replaced it. This messed up some reasonable invariants; for example,
that edges produced from one call-statement can easily grouped together
because they have the same location.

Instead, we now use a new algorithm that:
 - only removes "uninteresting" edges, thus getting rid of the problem
 of "synthesizing" in one edge the information from multiple edges, and
 thus maintaining reasonable invariants like the one mentioned above;
 - guarantees that neither the number of nodes nor the number of edges
 goes up; and
 - does NOT guarantee that ALL nodes corresponding to temporary
 variables are removed (the goal now is "reduce size" rather than
 "eliminate ALL temporaries")

Reviewed By: thizanne

Differential Revision: D37238743

fbshipit-source-id: 75406dac6cd524acfeeef10b116150f70e0c8dfd