[scripts][clang] redownload llvm tarball once on failure
Summary:
Whenever clang gets updated the checksum changes and unless there had
been some manual cleanup at some point, the previous llvm tarball will
still be around and make the script abort. Same goes if the previous
download was interrupted or failed for whatever reason.

When we detect a checksum mismatch, delete the offending llvm tarball.
This makes the rest of the script attempt to download it. If checksum
still fails after that, let it fail the script as before.

Reviewed By: martintrojer

Differential Revision: D37683072

fbshipit-source-id: f9ecb5b78cf85b9c4e89f30f527a5c839903d916