[sledge] Consider call/return structure for distance heuristics
Summary:
Add some richer structure to Sparse_trace goals for goal-directed analysis,
and enrich the distance heuristics computed by Llair.compute_distances
accordingly.  Currently, this is specialized to traces consisting of a source
and sink trace (separated by "__sledge_trace_separator__" in the trace input)
which are interpreted as a sequence of calls along the source trace, returns
back to the root, then calls along the sink trace.

Reviewed By: jberdine

Differential Revision: D37750638

fbshipit-source-id: 85e78c436bcf3e5c1d7f73eb146564f714f2eb46