Add a model for ordering erlang comparisons that support more types
Summary:
We add a model for the ordering comparisons in erlang (>, >=, =<, <), and move from a CFG-based
approach that assumes integers to a model that supports other types and is able to precisely compare
some of them.

Namely, the current model is able to precisely know that integers are always "lesser" than other
types, and when comparing an integer to a value of an unknown type, it splits the case where this
latter value is an integer or not to build precise results in both cases.

Comparing values that are not integer is now supported, albeit imprecisely. Concretely, we move from
assuming that compared values are integers (and not continuing in that branch in the opposite case)
to not assuming this but have an unknown result (thus continuing in the branch with no information
from the comparison).

Facebook :

Tests during the development seem to show that this does not have a lot of impact, but that this
impact is positive (see CI runs). We add some latent fps due to Infer limitations on sat solving, but
they do not seem to translate into actual issues.

Reviewed By: rgrig

Differential Revision: D37604102

fbshipit-source-id: 0f12e6a6dff322f39d50e807abfcc0f9ef3d0bce