Use safer function names for erlang comparison functions
Summary:
Using the builtin function names, such as `erlang:'=='`, will likely turn out to be a creeping danger
when we add the model from inequality. Indeed those contain slashes in the names, making any code
generating filenames from functions risky.

Although there is currently no evidence that the problematic behaviour arises, we get back on
generating safer custom builtin names to be on the safe side.

As a remark, we do lose generating a precise translation for code such as `erlang:'=='(1, 2)` for
free. If we ever need to actually analyse such code, this will be a one-line addition (per function)
on the custom models matching.

Reviewed By: rgrig

Differential Revision: D37758248

fbshipit-source-id: e878b9a34ea360b9f57a6e766f0e0d3d6b85eda5