Add a model for /= and =/= erlang comparisons
Summary:
This model also supports non-integer types, with the same precision as the other comparisons. Some
related tests are added.

Both functions are currently modelled as exactly_not_equal (that is, comparisons between "morally
equal" floats and integers will always be false). This does not change the current
behaviour. However, the new model would make it easier to leverage this limitation (or possibly to
return unknown result on float vs. integer comparisons).

Reviewed By: rgrig

Differential Revision: D37724030

fbshipit-source-id: 934f77e6c9ec8c3b29bb95e04574d46d28722570