[clang][frontend] Ensure only last value in a compound statement is saved
Summary:
For an assignment using a compound statement, e.g.:
```
Object* obj = ({
  do_stuff();
  do_other_stuff();
  result;
})
```
The final value assigned to `obj` should be `result`. However, when translating compound statements this was not
respected - instead, each statement in turn was assigned to `obj`. Even worse, these were processed in reverse order.
This commit ensures the last value in the sequence of statements is used.

Reviewed By: jvillard

Differential Revision: D37924742

fbshipit-source-id: 7eb10dfc129817525d25070c6dabd47b49be60f6