[resources][java] unknown call should remove allocated attribute
Summary:
When a resource API is not modeled yet, unknown calls on resources may generate FP because we lose the potential resource-closing that may occur during this call, or later (via delegation).
Currently the unknown calls do not havoc on Java programs.

We solve this issue by removing all `Allocated` attribute on objects that are reachable from unknown call arguments.

Reviewed By: ezgicicek

Differential Revision: D38068077

fbshipit-source-id: 3cad83c86714fd4b7a4f1fa6732baa3234ae0eea