[sledge] Pass scope to Frontend.add_sym
Summary:
Pass the "scope" (enclosing function or module) to add_sym when
computing debug locations, etc. instead of looking the scope up from
each value individually. It is clear how to safely compute the scope
when calling add_sym, while it is not possible to safely compute the
enclosing scope for an arbitrary llvalue.

Reviewed By: bennostein

Differential Revision: D38540963

fbshipit-source-id: db4d10b9d693e81978bce238f1d0dbd1f2245ed5