[datalog] Change fact format for procedure names
Summary:
This diff changes the `call_proc` field in `VirtualCall` to be a procedure name without the class information. That is because the actual class is known only if the call is static. To do this, we used a different pretty printer `unique_proc_id` for the procedure names.

The printer has the effect of dropping the "`Package.Class.`" prefix for virtual calls. Another change is that all calls with a `void` return type now have a suffix `:void`.

Reviewed By: bennostein

Differential Revision: D38538918

fbshipit-source-id: 758f85ba50a57aeaa7bd73880138aa8faa773fb9