[pulse] remove Taint.Kind hashconsing
Summary:
This didn't work because kinds get serialised into summaries then
deserialised, and deserialising kinds breaks hash consing as we don't
look up what their known canonical representative is at the point of
deserialisation. This is fixable but there's actually no datapoint that
this optimisation is necessary at the moment so get rid of it for now.

Reviewed By: skcho

Differential Revision: D38629213

fbshipit-source-id: 84c393ec45d12def8049750d302a4f75b3edf4e2