[pulse] fold concrete intervals inside the main formula domain
Summary:
It makes more sense for all the arithmetic domains to live inside the
formula. This way they can exchange information. For now, just copy the
interval domain inside the formula without taking advantage of this to
make it communicate equalities that it discovers, and without deleting
its copy in PathCondition.

Reviewed By: skcho

Differential Revision: D38501997

fbshipit-source-id: cf51addb41d02db594f165f3bffb87c02de6d648