[pulse] propagate Unsat eagerly
Summary:
Previously an abstract state could be unsatisfiable but the
unsatisfiability could be hidden in the path condition, which kept an
`is_unsat` boolean flag. When that flag was true, we can keep operating
on the state, and at some point we'll realise that it is unsat and
discard it. This could in theory cause unnecessary work.

Instead, always handle Unsat as soon as we detect it. This complicates
the result monad as some operations can introduce Unsat but not all. The
new interface in PulseOperationResult is hopefully clear enough.

This is the last step to completely get rid of PulsePathCondition.ml and
use PulseFormula.ml directly instead.

This revealed two issues with erlang models whereby there was a case
split that could be short-circuited if only one of the branches
failed/became Unsat.

Reviewed By: rgrig

Differential Revision: D38580369

fbshipit-source-id: d55e9d41f8cc11581b1a5d049247f2d6304567f1