[ConfigImpact] Handle Kotlin lazy method
Summary:
A function call to Kotlin's lazy method `C.foo` is translated to a function call to a getter
`C.getFoo`.

- The getter `C.getFoo` calls a magic function `Lazy.getValue` in it.
- The method `C$foo$2.invoke` contains the original function body.

`C$foo$2.invoke` is called somehow by `Lazy.getValue`, but the checker could not follow the
semantics before. To avoid the missing semantics, this diff it replaces the function call of
`Lazy.getValue` to `invoke`.

Reviewed By: ngorogiannis

Differential Revision: D38785182

fbshipit-source-id: 65d5e8e9d6f0c064d3db4751810b78eeab6fafe1