[pulse][alias] closure specialization now updates aliases' pvars
Summary:
commit 357daeed was a quickfix to work around the issue. The current
commit actually fixes the issue.
The problem was that if the closure specialization happens after an
alias specialization, then the pvars stored by the alias
specialization are not well localized: their procname is the alias
specialized version without the closure specialization meaning that when
the alias specialization is applied we try to find variables that
do not exist in the current one.

Reviewed By: da319

Differential Revision: D38978030

fbshipit-source-id: e85c6d232ad24437f41aa4f3a0b32909048f3796