[pulse][unnecessary copy] Track copies originating from source objects
Summary:
Previously, we were only tracking copies into either program variables occuring in the code or member fields.

However, there are other ways of creating copies that are worth tracking: e.g.  specifically copies from program variables (that are occuring in the code)  into temporaries as in

```

int get_size(A a) { return a.vec.size(); }

void unnecessary_intermetidate_copy(const A& my_a) {
  // we are copying the argument my_a unnecessarily because get_size doesn't
  // have const A&
  get_size(my_a);
}
```

This diff creates a new issue type for such issues.

With this diff, we are able to find copies at the callsites of const-refable functions which was not possible before.

AFAICS, this occurs only for copy constructors and we don't copy assign into temporary objects. I might be wrong so will see if CI detects new cases.

Reviewed By: skcho

Differential Revision: D39136719

fbshipit-source-id: 027ba56370af455d4f5e6135bd4237d007822bc2