[infer-erl][3/n] Validate type specs: ensure overloaded function specs have same arity
Summary:
We add a new case to `validate_form` that examines a `Spec` form, and ensures that all the overloads have the same arity (equal to that of the `Spec`, by checking that the length of the `argument` list of each spec disjunct is equal to the `arity` of the `function_` in the `Spec`).
Note that this does not check whether the `Function` corresponding to the `Spec` has the same arity as the `Spec` (or whether it even exists).

Differential Revision: D38355391

fbshipit-source-id: ec9e68cdd8cb26c55e6323725c94d9bd86ca63c4